Update LICENSE just in case there is any Gaggiuino code in there.
Attribution: If this project contains any Gaggiuino code, I hereby give attribution to Zer0-bit © Gaggiuino.

If that is the case, it should be removed, and then this license should be updated to GNU General Public License v2.0.
The future goal of this project is to only have original or licensed code, and then become GNU General Public License v2.0 after that is verified.

Anyone working this code must agree to not use any Gaggiuino code, and to replace any that they come across. Once it is verified that this is all original or compliant code, then change the license to to GNU General Public License v2.0.